reorganized mod information file structure
